export module for getting builder data
