+++ src/components/index.js
@@ -1,11 +1,11 @@
 import AppletSchemaBuilder from './ProtocolBuilder/Builder';
 
-import { html5Media } from 'markdown-it-html5-media';
 import appletBuilderStore from '../store/modules/appletBuilder';
 import config from '../config';
 
 import html5Embed from 'markdown-it-html5-embed';
 import markdownItImSize from 'markdown-it-imsize';
+import Protocol from '../models/Protocol';
 
 const Components = {
   AppletSchemaBuilder
@@ -27,4 +27,5 @@ export default {
       Vue.component(name, Components[name]);
     });
   },
+  getBuilderFormat: Protocol.getBuilderFormat,
 };
