+++ src/models/Protocol.js
@@ -1,8 +1,8 @@
 import api from '../utilities/api';
 import util from '../utilities/util';
 import Activity from './Activity';
+import Item from './Item';
 import _ from 'lodash';
-import { version } from 'core-js';
 
 export default class Protocol {
   constructor() {
@@ -141,12 +141,12 @@ export default class Protocol {
         updated: (field) => `Applet image was changed to ${_.get(newValue, field)}`,
         removed: (field) => `Applet image was removed`,
         inserted: (field) => `Applet image was added (${_.get(newValue, field)})`
-      }, 
+      },
       'schema:description': {
         updated: (field) => `Applet description was changed to ${_.get(newValue, field)}`,
         removed: (field) => `Applet description was removed`,
         inserted: (field) => `Applet description was added (${_.get(newValue, field)})`
-      }, 
+      },
     }
   }
 
@@ -157,7 +157,7 @@ export default class Protocol {
     } = old.protocol;
 
     const {
-      "data": currentData, 
+      "data": currentData,
       "activities": currentActivities
     } = current.protocol;
 
@@ -165,7 +165,7 @@ export default class Protocol {
     let versionUpgrade = '0.0.0';
     let updates = { activities: {} };
     let removed = {activities: [], items: []};
-  
+
     const metaInfoChanges = util.compareValues(oldData, currentData, Object.keys(logTemplates));
     const activityChanges = util.compareIDs(oldActivities, currentActivities, 'data._id');
 
@@ -306,6 +306,81 @@ export default class Protocol {
     }
 
     return protocolInfo;
+  }
+
+  static async parseApplet(data) {
+    const { applet, activities, items, protocol } = data;
+
+    let initialStoreData = {
+      ... await Protocol.parseJSONLD(applet, protocol),
+      valid: true,
+      activities: [],
+      tokenPrizeModal: false,
+    };
+
+    const activityModel = new Activity();
+    const itemModel = new Item();
+
+    Object.values(activities).forEach((act) => {
+      const activityInfo = Activity.parseJSONLD(act)
+      const activityItems = activityInfo.orderList.filter(key => items[key]).map((key) => {
+        return itemModel.getItemBuilderData(Item.parseJSONLD(items[key]));
+      });
+
+      const builderData = activityModel.getActivityBuilderData({
+        ...activityInfo,
+        items: activityItems,
+      });
+      builderData.index = initialStoreData.activities.length;
+
+      initialStoreData.activities.push(builderData);
+    });
+
+    initialStoreData.prizeActivity = initialStoreData.activities.find(activity => activity.isPrize);
+
+    return initialStoreData;
+  }
+
+  static formattedProtocol(protocol) {
+    const protocolModel = new Protocol();
+    const activities = protocol.activities.map(activity => {
+      const activityModel = new Activity();
+
+      activityModel.updateReferenceObject({
+        ...activity,
+        items: activity.items.map(item => {
+          const itemModel = new Item();
+
+          itemModel.updateReferenceObject(item);
+          return itemModel.getItemData()
+        })
+      });
+
+      return activityModel.getActivityData();
+    })
+
+    protocolModel.updateReferenceObject({
+      ...protocol,
+      activities
+    });
+
+    return protocolModel.getProtocolData();
+  }
+
+  static async getBuilderFormat(applet, upgradeVersion=false) {
+    const protocol = await this.parseApplet(applet)
+    const builderData = this.formattedProtocol(protocol);
+
+    if (upgradeVersion) {
+      builderData.protocol.data[
+        'schema:schemaVersion'
+      ] = original.protocol.data['schema:version'] = util.upgradeVersion(
+        original.protocol.data['schema:version'],
+        '0.0.1'
+      );
+    }
+
+    return builderData;
   }
 
   static checkValidation(protocol) {
