+++ src/store/modules/appletBuilder/getters.js
@@ -26,33 +26,10 @@ const activityGetters = {
 
 export default {
   ...activityGetters,
-  
+
   /** get formatted protocol data from current protocol state */
   formattedProtocol (state) {
-    const protocol = state.protocol;
-    const protocolModel = new Protocol();
-    const activities = protocol.activities.map(activity => {
-      const activityModel = new Activity();
-
-      activityModel.updateReferenceObject({
-        ...activity,
-        items: activity.items.map(item => {
-          const itemModel = new Item();
-
-          itemModel.updateReferenceObject(item);
-          return itemModel.getItemData()
-        })
-      });
-
-      return activityModel.getActivityData();
-    })
-
-    protocolModel.updateReferenceObject({
-      ...protocol,
-      activities
-    });
-
-    return protocolModel.getProtocolData();
+    return Protocol.formattedProtocol(state.protocol);
   },
 
   protocol (state) {
